+++ src/lib/cards/OutlineCardsSideBySide.svelte
@@ -1,8 +1,9 @@
-<script>
+<script lang="ts">
+	import type { OutlineObject } from '../../data/interfaces/interfaces';
 	import OutlineCardAnimated from './OutlineCardAnimated.svelte';
 
-	export let leftOutline;
-	export let rightOutline;
+	export let leftOutline: OutlineObject;
+	export let rightOutline: OutlineObject;
 </script>
 
 <div style="display: flex; gap: 5%; margin: 50px auto;">
