+++ src/lib/disemvoweler/wordDisemvoweler.svelte
@@ -1,7 +1,7 @@
-<script>
-	import { disemvowelWord } from '../../scripts/disemvowel.js';
-	import { randomIntFromInterval } from '../../scripts/helpers.ts';
-	export let word;
+<script lang="ts">
+	import { disemvowelWord } from '../../scripts/disemvowel';
+	import { randomIntFromInterval } from '../../scripts/helpers';
+	export let word: string;
 </script>
 
 <div class="disemvoweler-container" style="transform: rotate({randomIntFromInterval(-1, 1)}deg);">
